requirements.txt
Django~=5.1.1
django-ninja